More MSVC warning fixes

+++ PPP/GUI_private.h
@@ -27,7 +27,7 @@ namespace Graph_lib {
         }
         void detach(Shape& s)
         {
-            for (unsigned int i = shapes.size(); 0<i; --i)	// guess last attached will be first released
+            for (size_t i = shapes.size(); 0<i; --i)	// guess last attached will be first released
                 if (shapes[i-1]==&s)
                     shapes.erase(shapes.begin()+(i-1));//&shapes[i-1]);
         }
